Refactor retry logic
